Entity in application fixed

+++ src/framework/image.cpp
@@ -389,9 +389,10 @@ void Image::DrawRect(int x, int y, int w, int h, const Color& borderColor, int b
 
 void Image::DrawLineDDA(int x0, int y0, int x1, int y1, const Color& c){
     
-    int dx = x1- x0;
-    int dy = y1 - y0;
-    int d = (std::max)(abs(dx), abs(dy));
+    float dx = x1- x0;
+    float dy = y1 - y0;
+    float d = (std::max)(abs(dx), abs(dy));
+
     for(int i = 0; i <= d; i++) {
         int x = x0 + i * dx/d;
         int y = y0 + i * dy/d;
@@ -427,9 +428,9 @@ void Image::DrawTriangle(const Vector2& p0, const Vector2& p1, const Vector2& p2
 }
 
 void Image::ScanLineDDA(int x0, int y0, int x1, int y1, std::vector<Cell>& table) {
-    int dx = x1 - x0;
-    int dy = y1 - y0;
-    int d = (std::max)(abs(dx), abs(dy));
+    float dx = x1 - x0;
+    float dy = y1 - y0;
+    float d = (std::max)(abs(dx), abs(dy));
 
     for (int i = 0; i <= d; i++) {
         int x = x0 + i * dx / d;
